@@ -11,25 +11,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.6",
-    "@angular/cdk": "^10.2.5",
-    "@angular/common": "~10.0.6",
+    "@angular/animations": "^10.2.1",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "^10.2.1",
     "@angular/compiler": "~10.0.6",
-    "@angular/core": "~10.0.6",
+    "@angular/core": "^10.2.1",
     "@angular/forms": "~10.0.6",
-    "@angular/material": "^10.2.5",
+    "@angular/material": "^10.2.7",
     "@angular/platform-browser": "~10.0.6",
     "@angular/platform-browser-dynamic": "~10.0.6",
     "@angular/router": "~10.0.6",
     "bootstrap": "^4.5.3",
     "font-awesome": "^4.7.0",
+    "hammerjs": "^2.0.8",
+    "ngx-toastr": "^13.1.0",
     "rxjs": "~6.5.5",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1000.5",
-    "@angular/cli": "^10.1.7",
+    "@angular/cli": "^10.2.0",
     "@angular/compiler-cli": "~10.0.6",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
