@@ -3,14 +3,23 @@ import { OrdersComponent } from './orders/orders.component';
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-const routes: Routes = [
-  {path:'',redirectTo:'order',pathMatch:'full'},
-  {path:'orders',component:OrdersComponent},
-  {path:'order',children:[
-    {path:'',component:OrderComponent},
-    {path:'edit:/id',component:OrderComponent}
-  ]}
-];
+const routes: Routes =
+  [
+    { path: '', redirectTo: 'order', pathMatch: 'full' },
+
+    { path: 'orders', component: OrdersComponent },
+
+
+    // {path:'order',component:OrderComponent},
+    // {path:'edit/:id',component:OrderComponent}
+    {
+      path: 'order', children:
+        [
+          { path: '', component: OrderComponent },
+          { path: 'edit/:id', component: OrderComponent }
+        ]
+    }
+  ];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
